feat: generic plug-and-play purchasing system

+++ MyMusic.Common/Entities/Source.cs
@@ -4,16 +4,13 @@ namespace MyMusic.Common.Entities;
 
 public class Source
 {
-    public int Id { get; set; }
-    
-    [MaxLength(256)]
-    public required string Name { get; set; }
-    
-    [MaxLength(256)]
-    public required string Icon  { get; set; }
-    
-    [MaxLength(1024)]
-    public required string Address { get; set; }
-    
+    public long Id { get; set; }
+
+    [MaxLength(256)] public required string Name { get; set; }
+
+    [MaxLength(256)] public required string Icon { get; set; }
+
+    [MaxLength(1024)] public required string Address { get; set; }
+
     public required bool IsPaid { get; set; }
 }
